Rename TBANS tests folder
