+++ tests/tbans_tests/models/notifications/test_ping.py
