+++ tests/tbans_tests/models/notifications/test_update_favorites.py
