+++ tests/tbans_tests/models/subscriptions/test_subscription_batch_request.py
