+++ tests/tbans_tests/models/subscriptions/test_subscription_batch_response.py
@@ -2,7 +2,7 @@ import unittest2
 
 from tbans.models.subscriptions.subscription_batch_response import SubscriptionBatchResponse
 
-from tests.tbans.mocks.mock_response import MockResponse
+from tests.tbans_tests.mocks.mock_response import MockResponse
 
 
 class TestSubscriptionBatchResponse(unittest2.TestCase):
